Rearrange code; add Makefile; add install docs to readme
Move source code into a more standard layout and provide a Makefile
so that (re)compilation is a bit easier.  Add a 'bin' directory into
which to place executables and a 'obj' directory for object and
module files.

To build the codes, in the root directory do

    make -C src

Build targets include 'serial', 'openmp' and 'mpi' to only build the
codes using a particular parallelism.

Add this information to the README.md.